diplom/diplom_retail_network/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-rest-passwordreset==1.3.0
requests==2.31.0
ujson==5.9.0
PyYAML>=6.0.1
celery==5.3.4
redis==5.0.1
flower==2.0.1
